dom/base/test/test_messagePort.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=912456
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 912456 - port cloning</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=912456">Mozilla Bug 912456</a>
<script type="application/javascript">

  function testTransfer() {
    var a = new MessageChannel();
    ok(a, "MessageChannel created");

    window.addEventListener('message', receiveMessage, false);
    function receiveMessage(evt) {
      ok(evt.data.port, "Port has been received!");

      var a = new MessageChannel();
      ok(a, "MessageChannel created");

      try {
        evt.data.port.postMessage({port: a.port2});
        ok(false, "PostMessage should throw! - no transfered port");
      } catch(e) {
        ok(true, "PostMessage should throw! - no transfered port");
      }

      try {
        evt.data.port.postMessage({port: a.port2}, [a.port2, a.port2]);
        ok(false, "PostMessage should throw - no duplicate!");
      } catch(e) {
        ok(true, "PostMessage should throw - no duplicate!");
      }

      evt.data.port.postMessage({port: a.port2}, [a.port2]);
    }

    a.port1.onmessage = function(evt) {
      ok(evt.data.port, "Port has been received!");
      window.removeEventListener('message', receiveMessage);
      runTest();
    }

    try {
      postMessage({ port: a.port2}, 42, '*');
      ok(false, "PostMessage should throw! - no transfered port");
    } catch(e) {
      ok(true, "PostMessage should throw! - no transfered port");
    }

    try {
      postMessage({ port: a.port2}, 42, '*', [a.port2, a.port2]);
      ok(false, "PostMessage should throw - no duplicate!");
    } catch(e) {
      ok(true, "PostMessage should throw - no duplicate!");
    }

    postMessage({port: a.port2}, '*', [a.port2]);
  }

  function testPorts() {
    var a = new MessageChannel();
    ok(a, "MessageChannel created");

    window.addEventListener('message', receiveMessage, false);
    function receiveMessage(evt) {
      ok(evt.data, "Data is 42");
      ok(evt.ports, "Port is received");
      is(evt.ports.length, 1, "Ports.length is 1");

      var a = new MessageChannel();
      ok(a, "MessageChannel created");

      evt.ports[0].postMessage(42, [a.port2]);
    }

    a.port1.onmessage = function(evt) {
      ok(evt.data, "Data is 42");
      ok(evt.ports, "Port is received");
      is(evt.ports.length, 1, "Ports.length is 1");
      window.removeEventListener('message', receiveMessage);
      runTest();
    }

    postMessage(42, '*', [a.port2]);
  }

  var tests = [
    testTransfer,
    testPorts
  ];

  function runTest() {
    if (!tests.length) {
      SimpleTest.finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [["dom.messageChannel.enabled", true]]}, runTest);

</script>
</body>
</html>
